app/services/compliance/monitor.py
try:
    import requests
    from bs4 import BeautifulSoup
    from transformers import pipeline
except ImportError:
    requests = None
    BeautifulSoup = None
    pipeline = None

from typing import Dict, List, Any
import time

class ComplianceMonitor:
    def __init__(self):
        try:
            # Load summarization model from Hugging Face
            self.summarizer = pipeline("summarization", model="sshleifer/distilbart-cnn-12-6")
        except:
            self.summarizer = None

    async def check_updates(self) -> Dict[str, Any]:
        """
        Scrape regulatory sites and summarize new updates.
        """
        # Mock scraping results (in real app, use BeautifulSoup on actual URLs)
        raw_updates = [
            {
                "title": "New Anti-Money Laundering Regulations 2024",
                "content": "The central bank has issued new guidelines regarding the reporting of large transactions. Banks must now report any transaction over $5,000 within 24 hours. Failure to comply will result in significant fines and potential license suspension.",
                "source": "Central Bank Hub"
            },
            {
                "title": "Digital Asset Custody Framework",
                "content": "A new framework for the custody of digital assets has been proposed. Institutions holding crypto-assets on behalf of clients must maintain a minimum capital reserve and implement multi-sig architecture for cold storage.",
                "source": "FinReg Authority"
            }
        ]

        processed_updates = []
        for update in raw_updates:
            # Use AI to summarize if model is loaded
            if self.summarizer:
                summary = self.summarizer(update["content"], max_length=30, min_length=10, do_sample=False)[0]['summary_text']
            else:
                summary = update["content"][:100] + "..."

            processed_updates.append({
                "title": update["title"],
                "source": update["source"],
                "summary": summary,
                "risk_level": "High" if "fines" in update["content"].lower() else "Medium",
                "date": time.strftime("%Y-%m-%d")
            })

        return {
            "new_updates": processed_updates,
            "total_active_alerts": len(processed_updates)
        }

compliance_monitor = ComplianceMonitor()
